Fix AVR warnings; FAT FS needs to use off_t instead of size_t
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@3728 42af7a65-404d-4744-a932-0658087f49c3

+++ drivers/bch/bchdev_unregister.c
@@ -110,7 +110,7 @@ int bchdev_unregister(const char *chardev)
    * will hold a reference count on the state structure.
    */
 
-  ret = ioctl(fd, DIOC_GETPRIV, (unsigned long)&bch);
+  ret = ioctl(fd, DIOC_GETPRIV, (unsigned long)((uintptr_t)&bch));
   (void)close(fd);
 
   if (ret < 0)
